Price: allow currency to be a string value
This will allow every possible currency to be used as a currency. This is necessary because of Flow types since there is enum with "possible future values". To use it properly it's necessary to use wrapping component with Relay fragment and proper Flow typing. This is better design anyway because Price component is responsible only for the displaying and PaymentPrice component is responsible for data fetching.